requirements.txt
openai
python-dotenv
yt-dlp
streamlit
openai-whisper
setuptools-rust
